+++ nens_auth_client/signals.py
@@ -1,0 +1,3 @@
+import django.dispatch
+
+invitation_accepted = django.dispatch.Signal()